fix: reduce minimum accepted height
With testing manually with 3 different clusters of nodes, 1. api.chainweb.com, 2. chainweb.kaddex.com, 3. kadena2.runonflux.io, we've identified that the current kadena height check is too strict for any of those nodes to be considered successful. The new change is actively being used with Kadefi Chainweb nodes to see if they are synced and is the appropriate number for all chainweb clusters